refactor: move hilbert-r-tree from vendor/ to utils/
